refactor(admin): break circular dependencies in admin bundle
The admin bundle had two import cycles that rollup warns about on
build:admin. dialog.js and hotkeys.js imported each other because the
"?" help-key handler lived in hotkeys but needed openDialog; it is
app-level policy rather than a core hotkeys concern, so it moves to the
initializer which already wires up document-level keys. element_editor.js
and its delete_element_button child formed the second cycle through the
shared dispatchPageDirtyEvent helper, which now lives in its own utils
module so neither side has to import the other.
