+++ app/javascript/alchemy_admin/utils/dispatch_page_dirty_event.js
@@ -1,0 +1,7 @@
+export function dispatchPageDirtyEvent(data) {
+  document.dispatchEvent(
+    new CustomEvent("alchemy:page-dirty", {
+      detail: { tooltip: data.publishButtonTooltip }
+    })
+  )
+}
